+++ feature/client/src/commonMain/kotlin/com/mifos/feature/client/navigation/ClientNavigation.kt
@@ -17,9 +17,8 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
 import androidx.navigation.navArgument
 import com.mifos.core.common.utils.Constants
-import com.mifos.feature.client.NewFixedDepositAccount.FixedAccountdestination
-import com.mifos.feature.client.NewFixedDepositAccount.FixedDepositRoute
-import com.mifos.feature.client.NewFixedDepositAccount.navigateToNewFixedDepositRoute
+import com.mifos.feature.client.newFixedDepositAccount.fixedAccountdestination
+import com.mifos.feature.client.newFixedDepositAccount.navigateToNewFixedDepositRoute
 import com.mifos.feature.client.clientAddDocuments.AddDocumentRoute
 import com.mifos.feature.client.clientAddDocuments.clientAddDocumentGraphRoute
 import com.mifos.feature.client.clientAddDocuments.navigateToClientAddDocumentRoute
@@ -341,7 +340,7 @@ fun NavGraphBuilder.clientNavGraph(
         )
 
         shareAccountDestination()
-        FixedAccountdestination()
+        fixedAccountdestination()
     }
 }
