+++ feature/client/src/commonMain/kotlin/com/mifos/feature/client/newFixedDepositAccount/pages/DetailsPage.kt
@@ -1,8 +1,7 @@
-package com.mifos.feature.client.NewFixedDepositAccount.Pages
-
+package com.mifos.feature.client.newFixedDepositAccount.pages
 
 import androidclient.feature.client.generated.resources.Res
-import androidclient.feature.client.generated.resources.step_interest
+import androidclient.feature.client.generated.resources.step_details
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.height
@@ -15,9 +14,9 @@ import androidx.compose.ui.unit.dp
 import org.jetbrains.compose.resources.stringResource
 
 @Composable
-fun InterestPage(onNext: () -> Unit) {
+fun DetailsPage(onNext: () -> Unit) {
     Column(horizontalAlignment = Alignment.CenterHorizontally) {
-        Text(stringResource(Res.string.step_interest))
+        Text(stringResource(Res.string.step_details))
         Spacer(Modifier.height(8.dp))
         Button(onClick = onNext) {
             Text("Next Button")
